apps/api/tests/requirements-e2e.txt
# Minimal dependencies for E2E API tests
# These tests only make HTTP calls to the deployed API

# HTTP client
httpx>=0.27.0

# Test framework
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-timeout>=2.3.0

# JWT token generation
pyjwt>=2.8.0

# Convex client for database operations (migrated from Supabase)
# The Convex client is included in the main backend dependencies
# For E2E tests, mock-based testing is used via conftest.py fixtures

# Environment variables
python-dotenv>=1.0.0
